@Yanrin 2022-06-22
Исходный бенч:
```
goos: linux
goarch: amd64
pkg: github.com/fixme_my_friend/hw10_program_optimization
cpu: Intel(R) Core(TM) i5-7200U CPU @ 2.50GHz
BenchmarkTimeMemory-4   	    1940	   6024999 ns/op	  166922 B/op	    1628 allocs/op
```
Финальный бенч:
```
goos: linux
goarch: amd64
pkg: github.com/fixme_my_friend/hw10_program_optimization
cpu: Intel(R) Core(TM) i5-7200U CPU @ 2.50GHz
BenchmarkTimeMemory-4   	12470428	       945.6 ns/op	    4152 B/op	       3 allocs/op
```
Сравнение:
```
name          old time/op    new time/op    delta
TimeMemory-4    6.02ms ± 0%    0.00ms ± 0%  -99.98%

name          old alloc/op   new alloc/op   delta
TimeMemory-4     167kB ± 0%       4kB ± 0%  -97.51%

name          old allocs/op  new allocs/op  delta
TimeMemory-4     1.63k ± 0%     0.00k ± 0%  -99.82%

```